perf(manifoldpacks): width-size storage to pack slots (Serendipity convention)
Follow the IERC1155Serendipity Claim convention and width-size the
numeric fields so related storage packs into shared slots:

PackConfig (was 6 slots -> now 3):
  maxCardsSupply  uint256 -> uint32   (cap ~4.29e9 card units)
  cardsPerPack    uint256 -> uint16
  numberOfVariations uint256 -> uint8 (the <=255 cap is now a TYPE
                                       guarantee; drops the runtime check
                                       and the MAX_UINT_8 constant)
  ripStartDate    uint256 -> uint48
  ripEndDate      uint256 -> uint48
  -> the five numerics now share ONE slot; deliverBatch reads them in a
     single SLOAD instead of five.

Standalone storage:
  signer(address) + mintedCards(uint32) + ripSignatureRequired(bool)
    + transfersPaused(bool)  -> one slot (was 4)
  creatorContractAddress(address) + startingCardTokenId(uint80) -> one slot
  mintedCards accumulated in a local across a deliverBatch and written
    back once, keeping the warm signer slot hot.

Verified storage layout: config 96B (3 slots), signer-group in slot 19,
creator-group in slot 20. All 119 manifoldpacks tests green; size 23,184 B.
Deployed + retested end-to-end on Shape Sepolia (mint/pause/rip all pass);
getConfig decodes the packed struct correctly on-chain.

+++ packages/manifold/contracts/manifoldpacks/IManifoldPacksSeaDropShim.sol
@@ -10,22 +10,30 @@ interface IManifoldPacksSeaDropShim {
      * @notice Card-side configuration for the pack collection. Set once at
      *         `initializeCards` and owner-updatable via `updateConfig`.
      *
+     *         Numeric fields are width-sized (following the Serendipity `Claim`
+     *         convention: uint32 supply, uint48 dates, uint8 variations) so the
+     *         five of them pack into a SINGLE storage slot — one SLOAD on the
+     *         `deliverBatch` hot path instead of five.
+     *
      * @param maxCardsSupply    Optional hard cap on total card units minted
      *                          across all rips. `0` == unlimited. Cannot be set
      *                          below the already-minted count (`mintedCards`).
+     *                          `uint32` (max ~4.29e9 card units).
      * @param cardsPerPack      Exact number of card units a single pack yields
      *                          when ripped. Every `RipOrder`'s `amounts` MUST
      *                          sum to this value. Must be > 0. Freely updatable.
+     *                          `uint16` (max 65,535).
      * @param numberOfVariations Number of contiguous card variation tokenIds
-     *                          reserved on the cards core (`> 0`, `<= 255` —
-     *                          the uint8 variation cap). May be RAISED via
-     *                          `updateConfig` (reserving additional contiguous
-     *                          ids) but never lowered.
+     *                          reserved on the cards core (`> 0`). `uint8`, so
+     *                          the `<= 255` cap is now enforced by the type
+     *                          itself. May be RAISED via `updateConfig`
+     *                          (reserving additional contiguous ids) but never
+     *                          lowered.
      * @param ripStartDate      Earliest timestamp at which `deliverBatch` may
-     *                          rip (inclusive lower gate).
+     *                          rip (inclusive lower gate). `uint48`.
      * @param ripEndDate        Latest timestamp at which `deliverBatch` may rip.
      *                          `0` == no end. When non-zero must be strictly
-     *                          greater than `ripStartDate`.
+     *                          greater than `ripStartDate`. `uint48`.
      * @param cardsLocation     Folder-pattern base URI for card metadata. Card
      *                          `tokenURI` is `cardsLocation + (tokenId -
      *                          startingCardTokenId + 1)`. Freely updatable.
@@ -39,11 +47,11 @@ interface IManifoldPacksSeaDropShim {
      *                          built-in folder-pattern resolution.
      */
     struct PackConfig {
-        uint256 maxCardsSupply;
-        uint256 cardsPerPack;
-        uint256 numberOfVariations;
-        uint256 ripStartDate;
-        uint256 ripEndDate;
+        uint32 maxCardsSupply;
+        uint16 cardsPerPack;
+        uint8 numberOfVariations;
+        uint48 ripStartDate;
+        uint48 ripEndDate;
         string cardsLocation;
         address tokenURIExtension;
     }
